reformatted some code, attempted to fix angular acceleration again
