+++ src/Sensor.py
@@ -20,7 +20,7 @@ class Sensor:
     ANGULAR = 1
     RANGE = 2
     BOTH = 3
-    TOLERANCE = 0.15
+    TOLERANCE = 0.1
     WINDOW_WIDTH = 100
 
     def __init__(
